@@ -1,20 +1,20 @@
 name: vrchat_dart
-description: A starting point for Dart libraries or applications.
-version: 0.0.1
+description: An unofficial Dart library for VRChat API.
+version: 0.0.2
 homepage: https://github.com/GizmoOAO/vrchat_dart
 
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=2.12.1 <3.0.0'
 
 dependencies:
   dio: ^3.0.10
   cookie_jar: ^1.0.1
   dio_cookie_manager: ^1.0.0
-  json_annotation: ^3.1.1
+  json_annotation: ^4.0.0
   intl: ^0.17.0
 
 dev_dependencies:
   pedantic: ^1.9.2
   test: ^1.14.4
   build_runner: ^1.11.1
-  json_serializable: ^3.5.1
+  json_serializable: ^4.0.3
