@@ -13,16 +13,7 @@ Config _$ConfigFromJson(Map<String, dynamic> json) {
   );
 }
 
-Map<String, dynamic> _$ConfigToJson(Config instance) {
-  final val = <String, dynamic>{};
-
-  void writeNotNull(String key, dynamic value) {
-    if (value != null) {
-      val[key] = value;
-    }
-  }
-
-  writeNotNull('messageOfTheDay', instance.messageOfTheDay);
-  writeNotNull('clientApiKey', instance.clientApiKey);
-  return val;
-}
+Map<String, dynamic> _$ConfigToJson(Config instance) => <String, dynamic>{
+      'messageOfTheDay': instance.messageOfTheDay,
+      'clientApiKey': instance.clientApiKey,
+    };
