@@ -17,7 +17,7 @@ LimitedUser _$LimitedUserFromJson(Map<String, dynamic> json) {
     json['fallbackAvatar'] as String,
     json['userIcon'] as String,
     json['last_platform'] as String,
-    (json['tags'] as List)?.map((e) => e as String)?.toList(),
+    (json['tags'] as List<dynamic>).map((e) => e as String).toList(),
     json['developerType'] as String,
     json['status'] as String,
     json['statusDescription'] as String,
@@ -28,32 +28,23 @@ LimitedUser _$LimitedUserFromJson(Map<String, dynamic> json) {
   );
 }
 
-Map<String, dynamic> _$LimitedUserToJson(LimitedUser instance) {
-  final val = <String, dynamic>{};
-
-  void writeNotNull(String key, dynamic value) {
-    if (value != null) {
-      val[key] = value;
-    }
-  }
-
-  writeNotNull('id', instance.id);
-  writeNotNull('username', instance.username);
-  writeNotNull('displayName', instance.displayName);
-  writeNotNull('bio', instance.bio);
-  writeNotNull('currentAvatarImageUrl', instance.currentAvatarImageUrl);
-  writeNotNull('currentAvatarThumbnailImageUrl',
-      instance.currentAvatarThumbnailImageUrl);
-  writeNotNull('fallbackAvatar', instance.fallbackAvatar);
-  writeNotNull('userIcon', instance.userIcon);
-  writeNotNull('last_platform', instance.lastPlatform);
-  writeNotNull('tags', instance.tags);
-  writeNotNull('developerType', instance.developerType);
-  writeNotNull('status', instance.status);
-  writeNotNull('statusDescription', instance.statusDescription);
-  writeNotNull('friendKey', instance.friendKey);
-  writeNotNull('last_login', instance.lastLogin);
-  writeNotNull('isFriend', instance.isFriend);
-  writeNotNull('location', instance.location);
-  return val;
-}
+Map<String, dynamic> _$LimitedUserToJson(LimitedUser instance) =>
+    <String, dynamic>{
+      'id': instance.id,
+      'username': instance.username,
+      'displayName': instance.displayName,
+      'bio': instance.bio,
+      'currentAvatarImageUrl': instance.currentAvatarImageUrl,
+      'currentAvatarThumbnailImageUrl': instance.currentAvatarThumbnailImageUrl,
+      'fallbackAvatar': instance.fallbackAvatar,
+      'userIcon': instance.userIcon,
+      'last_platform': instance.lastPlatform,
+      'tags': instance.tags,
+      'developerType': instance.developerType,
+      'status': instance.status,
+      'statusDescription': instance.statusDescription,
+      'friendKey': instance.friendKey,
+      'last_login': instance.lastLogin,
+      'isFriend': instance.isFriend,
+      'location': instance.location,
+    };
